+++ nbproject/project.properties
@@ -59,7 +59,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=PatronLogin
+main.class=librarysystem.PatronLogin
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
