+++ src/librarysystem/LibrarianMainScreen.java
@@ -1,5 +1,13 @@
+package librarysystem;
 
-import java.awt.Color;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import javax.swing.JOptionPane;
 
 /*
  * To change this license header, choose License Headers in Project Properties.
@@ -7,19 +15,24 @@ import java.awt.Color;
  * and open the template in the editor.
  */
 
+
 /**
  *
  * @author Shanice Vergonia
  */
 public class LibrarianMainScreen extends javax.swing.JFrame {
-
+    private String patronId;
+    private static Connection con;
     /**
-     * Creates new form LibrarianMainScreen
+     * Creates new form PatronManScreen
      */
-    public LibrarianMainScreen() {
+    public LibrarianMainScreen(String patronId, Connection con) {
         initComponents();
-        welcome2_label.setForeground(Color.white);
-        this.setLocationRelativeTo(this);
+        
+        this.patronId = patronId;
+        this.con = con;
+        
+        this.setLocationRelativeTo(null);
     }
 
     /**
@@ -31,25 +44,29 @@ public class LibrarianMainScreen extends javax.swing.JFrame {
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
-        jSeparator1 = new javax.swing.JSeparator();
+        welcome1_label = new javax.swing.JLabel();
         manageTransacation_b = new javax.swing.JButton();
         manageUsers_b = new javax.swing.JButton();
         manageBooks_b = new javax.swing.JButton();
-        welcome2_label = new javax.swing.JLabel();
         jLabel2 = new javax.swing.JLabel();
         jMenuBar1 = new javax.swing.JMenuBar();
-        options_jmenu = new javax.swing.JMenu();
+        options1_jmenu = new javax.swing.JMenu();
         menu_logout = new javax.swing.JMenuItem();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setLocation(new java.awt.Point(0, 0));
-        setPreferredSize(new java.awt.Dimension(1000, 600));
+        setMaximumSize(null);
+        setMinimumSize(null);
+        setPreferredSize(new java.awt.Dimension(350, 540));
+        setResizable(false);
+        setSize(new java.awt.Dimension(0, 0));
         getContentPane().setLayout(null);
 
-        jSeparator1.setForeground(new java.awt.Color(255, 255, 255));
-        jSeparator1.setOrientation(javax.swing.SwingConstants.VERTICAL);
-        getContentPane().add(jSeparator1);
-        jSeparator1.setBounds(280, 10, 2, 400);
+        welcome1_label.setFont(new java.awt.Font("Nirmala UI", 1, 24)); // NOI18N
+        welcome1_label.setIcon(new javax.swing.ImageIcon(getClass().getResource("/welcome_librarian.png"))); // NOI18N
+        welcome1_label.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        getContentPane().add(welcome1_label);
+        welcome1_label.setBounds(40, 90, 250, 70);
 
         manageTransacation_b.setBackground(new java.awt.Color(255, 255, 255));
         manageTransacation_b.setFont(new java.awt.Font("Nirmala UI", 1, 12)); // NOI18N
@@ -61,34 +78,43 @@ public class LibrarianMainScreen extends javax.swing.JFrame {
             }
         });
         getContentPane().add(manageTransacation_b);
-        manageTransacation_b.setBounds(50, 180, 176, 40);
+        manageTransacation_b.setBounds(80, 180, 176, 40);
 
         manageUsers_b.setBackground(new java.awt.Color(255, 255, 255));
         manageUsers_b.setFont(new java.awt.Font("Nirmala UI", 1, 12)); // NOI18N
         manageUsers_b.setIcon(new javax.swing.ImageIcon(getClass().getResource("/gear.png"))); // NOI18N
         manageUsers_b.setText("Manage Users");
+        manageUsers_b.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                manageUsers_bActionPerformed(evt);
+            }
+        });
         getContentPane().add(manageUsers_b);
-        manageUsers_b.setBounds(50, 240, 170, 40);
+        manageUsers_b.setBounds(80, 240, 170, 40);
 
         manageBooks_b.setBackground(new java.awt.Color(255, 255, 255));
         manageBooks_b.setFont(new java.awt.Font("Nirmala UI", 1, 12)); // NOI18N
         manageBooks_b.setIcon(new javax.swing.ImageIcon(getClass().getResource("/books.png"))); // NOI18N
         manageBooks_b.setText("Manage Books");
+        manageBooks_b.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                manageBooks_bActionPerformed(evt);
+            }
+        });
         getContentPane().add(manageBooks_b);
-        manageBooks_b.setBounds(50, 300, 170, 38);
-
-        welcome2_label.setFont(new java.awt.Font("Nirmala UI", 1, 24)); // NOI18N
-        welcome2_label.setIcon(new javax.swing.ImageIcon(getClass().getResource("/welcome_librarian.png"))); // NOI18N
-        welcome2_label.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        getContentPane().add(welcome2_label);
-        welcome2_label.setBounds(20, 80, 250, 80);
+        manageBooks_b.setBounds(80, 300, 170, 38);
 
         jLabel2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/library-services4.jpg"))); // NOI18N
         jLabel2.setText("jLabel2");
         getContentPane().add(jLabel2);
-        jLabel2.setBounds(0, -250, 1050, 870);
+        jLabel2.setBounds(0, 0, 350, 540);
 
-        options_jmenu.setText("Options");
+        options1_jmenu.setText("Options");
+        options1_jmenu.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                options1_jmenuActionPerformed(evt);
+            }
+        });
 
         menu_logout.setText("Logout");
         menu_logout.addActionListener(new java.awt.event.ActionListener() {
@@ -96,71 +122,44 @@ public class LibrarianMainScreen extends javax.swing.JFrame {
                 menu_logoutActionPerformed(evt);
             }
         });
-        options_jmenu.add(menu_logout);
+        options1_jmenu.add(menu_logout);
 
-        jMenuBar1.add(options_jmenu);
+        jMenuBar1.add(options1_jmenu);
 
         setJMenuBar(jMenuBar1);
 
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
-    private void manageTransacation_bActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_manageTransacation_bActionPerformed
-        // TODO add your handling code here:
-        
-       // Statement ps = new con.createStatement();
-    }//GEN-LAST:event_manageTransacation_bActionPerformed
-
     private void menu_logoutActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_menu_logoutActionPerformed
-        // TODO add your handling code here:
-        new LibrarianLogin().setVisible(true);
+        new LibrarianLogin().setVisible(true); 
         this.dispose();
     }//GEN-LAST:event_menu_logoutActionPerformed
 
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(LibrarianMainScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(LibrarianMainScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(LibrarianMainScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(LibrarianMainScreen.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
+    private void options1_jmenuActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_options1_jmenuActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_options1_jmenuActionPerformed
 
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new LibrarianMainScreen().setVisible(true);
-            }
-        });
-    }
+    private void manageTransacation_bActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_manageTransacation_bActionPerformed
+
+    }//GEN-LAST:event_manageTransacation_bActionPerformed
+
+    private void manageUsers_bActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_manageUsers_bActionPerformed
+        new LibrarianManageUser(this, con).setVisible(true);
+    }//GEN-LAST:event_manageUsers_bActionPerformed
+
+    private void manageBooks_bActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_manageBooks_bActionPerformed
+        new EditBook(this, "", con).setVisible(true);
+    }//GEN-LAST:event_manageBooks_bActionPerformed
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JLabel jLabel2;
     private javax.swing.JMenuBar jMenuBar1;
-    private javax.swing.JSeparator jSeparator1;
     private javax.swing.JButton manageBooks_b;
     private javax.swing.JButton manageTransacation_b;
     private javax.swing.JButton manageUsers_b;
     private javax.swing.JMenuItem menu_logout;
-    private javax.swing.JMenu options_jmenu;
-    private javax.swing.JLabel welcome2_label;
+    private javax.swing.JMenu options1_jmenu;
+    private javax.swing.JLabel welcome1_label;
     // End of variables declaration//GEN-END:variables
 }
