+++ src/librarysystem/LibrarianManageUser.form
@@ -1,14 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 
-<Form version="1.3" maxVersion="1.9" type="org.netbeans.modules.form.forminfo.JDialogFormInfo">
+<Form version="1.5" maxVersion="1.9" type="org.netbeans.modules.form.forminfo.JDialogFormInfo">
   <Properties>
     <Property name="defaultCloseOperation" type="int" value="2"/>
-    <Property name="maximumSize" type="java.awt.Dimension" editor="org.netbeans.beaninfo.editors.DimensionEditor">
-      <Dimension value="null"/>
-    </Property>
-    <Property name="preferredSize" type="java.awt.Dimension" editor="org.netbeans.beaninfo.editors.DimensionEditor">
-      <Dimension value="[860, 530]"/>
-    </Property>
   </Properties>
   <SyntheticProperties>
     <SyntheticProperty name="formSizePolicy" type="int" value="1"/>
@@ -34,8 +28,8 @@
               <Group type="103" groupAlignment="0" attributes="0">
                   <Group type="102" attributes="0">
                       <Group type="103" groupAlignment="0" attributes="0">
-                          <Component id="searchTextField" alignment="0" min="-2" pref="176" max="-2" attributes="0"/>
-                          <Component id="searchHint" alignment="0" min="-2" pref="165" max="-2" attributes="0"/>
+                          <Component id="searchTextField" min="-2" pref="176" max="-2" attributes="0"/>
+                          <Component id="searchHint" min="-2" max="-2" attributes="0"/>
                       </Group>
                       <EmptySpace type="unrelated" max="-2" attributes="0"/>
                       <Component id="searchButton" min="-2" max="-2" attributes="0"/>
@@ -44,6 +38,8 @@
                       <Group type="102" attributes="0">
                           <Component id="returnButton" min="-2" pref="113" max="-2" attributes="0"/>
                           <EmptySpace max="32767" attributes="0"/>
+                          <Component id="addUserButton" min="-2" pref="102" max="-2" attributes="0"/>
+                          <EmptySpace type="unrelated" max="-2" attributes="0"/>
                           <Component id="editUserButton" min="-2" pref="102" max="-2" attributes="0"/>
                       </Group>
                       <Component id="scrollPane" alignment="0" min="-2" pref="746" max="-2" attributes="0"/>
@@ -69,6 +65,7 @@
               <Group type="103" groupAlignment="3" attributes="0">
                   <Component id="returnButton" alignment="3" min="-2" max="-2" attributes="0"/>
                   <Component id="editUserButton" alignment="3" min="-2" max="-2" attributes="0"/>
+                  <Component id="addUserButton" alignment="3" min="-2" max="-2" attributes="0"/>
               </Group>
               <EmptySpace pref="37" max="32767" attributes="0"/>
           </Group>
@@ -81,7 +78,7 @@
         <Property name="font" type="java.awt.Font" editor="org.netbeans.beaninfo.editors.FontEditor">
           <Font name="Nirmala UI" size="14" style="1"/>
         </Property>
-        <Property name="text" type="java.lang.String" value="Search for user"/>
+        <Property name="text" type="java.lang.String" value="Search for user by Login ID"/>
       </Properties>
     </Component>
     <Component class="javax.swing.JTextField" name="searchTextField">
@@ -155,6 +152,14 @@
       </Properties>
       <Events>
         <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="editUserButtonActionPerformed"/>
+      </Events>
+    </Component>
+    <Component class="javax.swing.JButton" name="addUserButton">
+      <Properties>
+        <Property name="text" type="java.lang.String" value="Add user"/>
+      </Properties>
+      <Events>
+        <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="addUserButtonActionPerformed"/>
       </Events>
     </Component>
   </SubComponents>
